# src/main/java/com/polarbookshop/orderservice/model/OrderRequest.java
package com.polarbookshop.orderservice.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.*;

@Data
@Generated
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class OrderRequest {

    @JsonProperty("isbn")
    @NotBlank(message = "The book ISBN must be defined.")
    private String isbn;

    @NotNull(message = "The book quantity must be defined.")
    @Min(value = 1, message = "You must order at least 1 item.")
    @Max(value = 5, message = "You cannot order more than 5 items.")
    @JsonProperty("quantity")
    private Integer quantity;
}