# src/main/resources/api.yml
openapi: 3.1.0
info:
  title: Order Service APIs
  description: "This lists all the Order Service API Calls. The Calls are OAuth2 secured,\
    \ so please use your client ID and Secret to test them out."
  version: v1.0
servers:
  - url: http://localhost:9002
paths:
  /api/v1/orders:
    get:
      tags:
        - order-controller
      summary: View all orders
      operationId: getAllOrders
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseModel"
        "403":
          description: The principal associated with the request does not have sufficient
            rights to preform this operation on the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "422":
          description: The request was syntactically correct but was not semantically
            correct. Usually indicating a validation problem.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "400":
          description: Indicates that the server could not understand the request.
            This is not the same as 422 which indicates a validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "401":
          description: Unauthorized.  This will be returned when no authentication
            information is provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseModel"
      security:
        - oAuth2ClientCredentials:
            - application.automotive-credit-requests.bcvc.create-credit-requests.read
    post:
      tags:
        - order-controller
      summary: Post an Order
      operationId: submitOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitOrderResponseModel"
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitOrderResponseModel"
        "403":
          description: The principal associated with the request does not have sufficient
            rights to preform this operation on the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "422":
          description: The request was syntactically correct but was not semantically
            correct. Usually indicating a validation problem.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "404":
          description: The requested resource was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "400":
          description: Indicates that the server could not understand the request.
            This is not the same as 422 which indicates a validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
        "401":
          description: Unauthorized.  This will be returned when no authentication
            information is provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyDataResponse"
      security:
        - oAuth2ClientCredentials:
            - application.automotive-credit-requests.bcvc.create-credit-requests.write
components:
  schemas:
    NotificationResponse:
      type: object
      properties:
        code:
          type: string
        source:
          type: string
        severity:
          type: string
        message:
          type: string
        uuid:
          type: string
          format: uuid
        notification_dt:
          type: string
          format: date-time
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        isbn:
          type: string
        title:
          type: string
        price:
          type: number
        quantity:
          type: integer
          format: int32
        status:
          type: string
          enum:
            - ACCEPTED
            - REJECTED
            - PENDING
            - DISPATCHED
        created_date:
          type: string
          format: date-time
    SubmitOrderResponseModel:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/OrderResponse"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationResponse"
    OrderRequest:
      type: object
      properties:
        isbn:
          type: string
        quantity:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
      required:
        - isbn
        - quantity
    EmptyDataResponse:
      type: object
      properties:
        data:
          type: string
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
    Notification:
      type: object
      properties:
        category:
          type: string
        code:
          type: string
        message:
          type: string
        description:
          type: string
        action:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: object
        uuid:
          type: string
        notification_dt:
          type: string
          format: date-time
        severity:
          type: string
        field_name:
          type: string
    OrderResponseModel:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponse"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationResponse"
  securitySchemes:
    oAuth2ClientCredentials:
      type: oauth2
      description: Oauth 2 Client Credential
      flows:
        clientCredentials:
          tokenUrl: http://localhost:9000/realms/Microservices-demo-realm/protocol/openid-connect/token
          scopes:
            application.automotive-credit-requests.bcvc.create-credit-requests.write: Create
              Credit Request
            application.automotive-credit-requests.bcvc.get-credit-requests.read: Retrieve
              Credit Request